[core] Moving all Actor Management related code in the worker to a separate bazel package. (#60505)
Pure refactoring. 

Moving all actor management code inside the worker to its own bazel
submodule.

In the final state, I'd like to have the owner-related actor management
code completely separated from the submitter-related actor client code.
