+++ src/ray/core_worker/actor_management/BUILD.bazel
@@ -1,0 +1,59 @@
+load("//bazel:ray.bzl", "ray_cc_library")
+
+ray_cc_library(
+    name = "actor_manager",
+    srcs = ["actor_manager.cc"],
+    hdrs = ["actor_manager.h"],
+    visibility = ["//src/ray/core_worker:__subpackages__"],
+    deps = [
+        ":actor_creator",
+        ":actor_handle",
+        "//src/ray/common:id",
+        "//src/ray/common:protobuf_utils",
+        "//src/ray/common:task_common",
+        "//src/ray/core_worker:common",
+        "//src/ray/core_worker:core_worker_context",
+        "//src/ray/core_worker:reference_counter_interface",
+        "//src/ray/core_worker/task_submission:actor_task_submitter",
+        "//src/ray/gcs_rpc_client:gcs_client",
+        "//src/ray/protobuf:core_worker_cc_proto",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_googletest//:gtest_prod",
+    ],
+)
+
+ray_cc_library(
+    name = "actor_handle",
+    srcs = ["actor_handle.cc"],
+    hdrs = ["actor_handle.h"],
+    visibility = ["//src/ray/core_worker:__subpackages__"],
+    deps = [
+        "//src/ray/common:id",
+        "//src/ray/common:task_common",
+        "//src/ray/core_worker:core_worker_context",
+        "//src/ray/protobuf:core_worker_cc_proto",
+        "//src/ray/protobuf:gcs_cc_proto",
+        "@com_google_absl//absl/types:optional",
+        "@com_google_googletest//:gtest_prod",
+    ],
+)
+
+ray_cc_library(
+    name = "actor_creator",
+    srcs = ["actor_creator.cc"],
+    hdrs = ["actor_creator.h"],
+    visibility = ["//src/ray/core_worker:__subpackages__"],
+    deps = [
+        "//src/ray/gcs_rpc_client:gcs_client",
+        "//src/ray/util:thread_utils",
+    ],
+)
+
+ray_cc_library(
+    name = "fake_actor_creator",
+    hdrs = ["fake_actor_creator.h"],
+    visibility = ["//src/ray/core_worker:__subpackages__"],
+    deps = [
+        ":actor_creator",
+    ],
+)
