+++ src/ray/core_worker/actor_management/tests/BUILD.bazel
@@ -1,0 +1,36 @@
+load("//bazel:ray.bzl", "ray_cc_test")
+
+ray_cc_test(
+    name = "actor_creator_test",
+    size = "small",
+    srcs = ["actor_creator_test.cc"],
+    tags = ["team:core"],
+    deps = [
+        "//:ray_mock",
+        "//src/ray/common:test_utils",
+        "//src/ray/core_worker/actor_management:actor_creator",
+        "//src/ray/gcs_rpc_client:gcs_client",
+        "//src/ray/util:path_utils",
+        "//src/ray/util:raii",
+        "@com_google_googletest//:gtest",
+        "@com_google_googletest//:gtest_main",
+    ],
+)
+
+ray_cc_test(
+    name = "actor_manager_test",
+    size = "small",
+    srcs = ["actor_manager_test.cc"],
+    tags = ["team:core"],
+    deps = [
+        "//:ray_mock",
+        "//src/ray/common:test_utils",
+        "//src/ray/core_worker:reference_counter",
+        "//src/ray/core_worker/actor_management:actor_manager",
+        "//src/ray/gcs_rpc_client:gcs_client",
+        "//src/ray/pubsub:fake_publisher",
+        "//src/ray/pubsub:fake_subscriber",
+        "@com_google_googletest//:gtest",
+        "@com_google_googletest//:gtest_main",
+    ],
+)
