+++ src/ray/core_worker/lib/java/io_ray_runtime_RayNativeRuntime.cc
@@ -26,7 +26,7 @@
 #include "jni_utils.h"  // NOLINT(build/include_subdir)
 #include "ray/common/id.h"
 #include "ray/common/ray_config.h"
-#include "ray/core_worker/actor_handle.h"
+#include "ray/core_worker/actor_management/actor_handle.h"
 #include "ray/core_worker/core_worker.h"
 #include "ray/util/time.h"
