+++ src/ray/core_worker/task_submission/tests/BUILD.bazel
@@ -9,7 +9,7 @@ ray_cc_test(
         "//:ray_mock",
         "//src/ray/common:task_common",
         "//src/ray/common:test_utils",
-        "//src/ray/core_worker:fake_actor_creator",
+        "//src/ray/core_worker/actor_management:fake_actor_creator",
         "//src/ray/core_worker/task_submission:dependency_resolver",
         "@com_google_googletest//:gtest",
     ],
@@ -53,10 +53,10 @@ ray_cc_test(
         "//src/ray/common:asio",
         "//src/ray/common:task_common",
         "//src/ray/common:test_utils",
-        "//src/ray/core_worker:actor_creator",
-        "//src/ray/core_worker:fake_actor_creator",
         "//src/ray/core_worker:reference_counter",
         "//src/ray/core_worker:task_manager",
+        "//src/ray/core_worker/actor_management:actor_creator",
+        "//src/ray/core_worker/actor_management:fake_actor_creator",
         "//src/ray/core_worker_rpc_client:fake_core_worker_client",
         "//src/ray/pubsub:fake_publisher",
         "//src/ray/pubsub:fake_subscriber",
@@ -75,8 +75,8 @@ ray_cc_test(
         "//:ray_mock",
         "//src/ray/common:task_common",
         "//src/ray/common:test_utils",
-        "//src/ray/core_worker:fake_actor_creator",
         "//src/ray/core_worker:memory_store",
+        "//src/ray/core_worker/actor_management:fake_actor_creator",
         "//src/ray/core_worker/task_submission:normal_task_submitter",
         "//src/ray/core_worker_rpc_client:core_worker_client_pool",
         "//src/ray/core_worker_rpc_client:fake_core_worker_client",
