+++ src/ray/core_worker/tests/BUILD.bazel
@@ -149,23 +149,6 @@ ray_cc_test(
 )
 
 ray_cc_test(
-    name = "actor_creator_test",
-    size = "small",
-    srcs = ["actor_creator_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//:ray_mock",
-        "//src/ray/common:test_utils",
-        "//src/ray/core_worker:actor_creator",
-        "//src/ray/gcs_rpc_client:gcs_client",
-        "//src/ray/util:path_utils",
-        "//src/ray/util:raii",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
     name = "generator_waiter_test",
     size = "small",
     srcs = ["generator_waiter_test.cc"],
@@ -176,24 +159,6 @@ ray_cc_test(
         "//src/ray/core_worker:common",
         "//src/ray/core_worker:generator_waiter",
         "//src/ray/gcs_rpc_client:gcs_client",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
-    name = "actor_manager_test",
-    size = "small",
-    srcs = ["actor_manager_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//:ray_mock",
-        "//src/ray/common:test_utils",
-        "//src/ray/core_worker:actor_manager",
-        "//src/ray/core_worker:reference_counter",
-        "//src/ray/gcs_rpc_client:gcs_client",
-        "//src/ray/pubsub:fake_publisher",
-        "//src/ray/pubsub:fake_subscriber",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
     ],
